add mintViewer

+++ src/interfaces/IETH20.sol
@@ -2,10 +2,10 @@
 pragma solidity =0.8.17;
 
 interface IETH20Events {
-    event Approval(address indexed src, address indexed guy, uint wad);
-    event Transfer(address indexed src, address indexed dst, uint wad);
-    event Deposit(address indexed dst, uint wad);
-    event Withdrawal(address indexed src, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 }
 
 interface IETH20 is IETH20Events {
@@ -16,20 +16,13 @@ interface IETH20 is IETH20Events {
 
     // WETH specific functions
     function deposit() external payable;
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 
     // ERC20 standard interface
-    function totalSupply() external view returns (uint);
-    function balanceOf(address owner) external view returns (uint);
-    function allowance(
-        address owner,
-        address spender
-    ) external view returns (uint);
-    function approve(address spender, uint value) external returns (bool);
-    function transfer(address to, uint value) external returns (bool);
-    function transferFrom(
-        address from,
-        address to,
-        uint value
-    ) external returns (bool);
+    function totalSupply() external view returns (uint256);
+    function balanceOf(address owner) external view returns (uint256);
+    function allowance(address owner, address spender) external view returns (uint256);
+    function approve(address spender, uint256 value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
+    function transferFrom(address from, address to, uint256 value) external returns (bool);
 }
